+++ network/src/main/java/ru/surfstudio/android/network/etag/storage/EtagCache.java
@@ -17,16 +17,16 @@ package ru.surfstudio.android.network.etag.storage;
 
 import javax.inject.Named;
 
-import ru.surfstudio.android.filestorage.BaseTextLocalCache;
 import ru.surfstudio.android.filestorage.CacheConstant;
 import ru.surfstudio.android.filestorage.naming.Sha256NamingProcessor;
-import ru.surfstudio.android.filestorage.processor.CacheFileProcessor;
+import ru.surfstudio.android.filestorage.processor.FileProcessor;
+import ru.surfstudio.android.filestorage.storage.BaseTextFileStorage;
 
-public final class EtagCache extends BaseTextLocalCache {
+final public class EtagCache extends BaseTextFileStorage {
 
     private static final int CACHE_SIZE = 5000;
 
-    public EtagCache(@Named(CacheConstant.NO_BACKUP_STORAGE_DIR_NAME) final String noBackupStorageDir) {
-        super(new CacheFileProcessor(noBackupStorageDir, "etag", CACHE_SIZE), new Sha256NamingProcessor());
+    public EtagCache(@Named(CacheConstant.INTERNAL_CACHE_DIR_DAGGER_NAME) final String cacheDir) {
+        super(new FileProcessor(cacheDir, "etag", CACHE_SIZE), new Sha256NamingProcessor());
     }
 }